removing seed migrations; using rake task
thinking i should use the rake task to load data.
in the future there should probably be a rake task that creates
a yml file for a season from the web and another task that loads
the db from the yml. will still need a way to detect changes in
game times