wrapping the season jobs in a transaction block
if the loading of the data results in too many games/teams,
throw an error and rollback the transaction.

this is a quick hack around the possibility of a game's time
changing. the change would result in the same game in the db
twice, once for each time. this would catch that there are too
many games and rollback. in the future, an email should go out
when this happens in order to provoke manual action