adding loading icons to controllers #35
note the incredibly hacky timeout in the angular controllers.
this is because the template was loading after the controller
finished its query. so if i put a loading icon, the call for the
icon would come after the icon for the data, thus not being able
to actually show the icon during the data download. defeating the
whole point.

i'm sure i'm doing something wrong with my angular setup. my next
attempt will be to combine the two calls into one, thus needing
only one icon and only one controller. this at minimum is more
intuitive and the cleaner approach. maybe doing it will give me
some insight in what's going on with my angular setup.