[cleanup] explictly declare start time
using `now` for the start time caused funkiness in the test on
codeship. but it is bad practice to do that anyway. removing it
seems to fix codeship.

also correctly using the last game in the list, which is an orphan.

also removing the development `current` tag that never should have
made it into git