filter games more reliably
it is more reliable (although perhaps more resource intensive)
to filter the games in ruby rather than filtering via xpath.

the xpath filter had the edge case of not properly doing a
`preceding-sibling` in the scenario where it was the first day
of a month. just grabing them all and then looping through afterward
will suffice for now.