refactor controller spec to focus on games
this decouples the game_controller spec from the logic of
determining tiers, which should be tested elsewhere. Now the
the game_controller spec focus on making sure, given a set of
tiers, that the correct games comes back. This makes more sense
since that's really all the logic the controller currently has.
It's all game-grade logic.

The test is now a lot simplier, particularly the setup

NOTE: For the most part, the tests covering the tier calculation
logic exists, but I'm missing metric-specific tests. That's a
TODO