sort calculated averages
This adds back the sorting that was originally added in d04e94f.
It was accidentally removed (along with its test) in 22a61e7.
This adds the sorting back.

Also includes a little rubo:cop:-inspired refactor of a method,
since sorting made it too big.

NOTE: The teams are sorted by their tier average. Before this big
feature of multiple metrics, they were sorted by their record.
Since there is only one metric currently, everybody in tier 1 will
have an average of 1.0. Thus the sorting within is indeterminate.
What that means for the UI is that tier 1 teams won't be sorted by
their record (which is what the UI actually shows). This means it
will look weird to see a team with a worse record on top of another
team, but this is technically accurate since they both have an
average of 1.0. There will be another ticket to redo the UI in
order not to imply that the tier is soley based on the record. but
I'm going to leave it like that for now and handle it on the other
ticket, which will probably get done right after I had the second
metric.